+++ src/app/dashboard/_components/sidepanel.tsx
@@ -6,7 +6,7 @@ import { SaleIcon } from "@/components/icons/saleicon";
 export default function SidePanel() {
   return (
     <>
-      <div className="pb-12 w-1/5">
+      <div className="pb-12 w-1/6 h-screen bg-white rounded-lg">
         <div className="space-y-4 py-4">
           <div className="px-3 py-2">
             <h2 className="mb-2 px-4 text-lg font-semibold tracking-tight">
@@ -16,7 +16,7 @@ export default function SidePanel() {
               <Button
                 asChild
                 variant="ghost"
-                className="hover:bg-slate-300 w-fit justify-start flex gap-1"
+                className="hover:bg-slate-300 w-fit justify-start flex gap-2 bg-white px-3"
               >
                 <Link href="">
                   <Box />
@@ -26,7 +26,7 @@ export default function SidePanel() {
               <Button
                 asChild
                 variant="ghost"
-                className="hover:bg-slate-300 w-fit justify-start flex gap-2"
+                className="hover:bg-slate-300 w-fit justify-start flex gap-2 bg-white px-3"
               >
                 <Link href="">
                   <SaleIcon />
